moving utils in sentence_topic_analysis.py to utils folder
